requirements.txt
fastapi==0.133.1
uvicorn==0.41.0
asyncpg==0.31.0
pgvector==0.4.2
python-dotenv==1.2.1
numpy==2.4.2
sentence-transformers==5.2.3
torch==2.10.0
ollama==0.6.1
python-multipart==0.0.22
pydantic-settings==2.13.1